requirements.txt
pyautogui
pyserial